Update Github_Tutorial.ino
Changed byte to int for myValue in Github_Tutorial.ino file.

+++ Github_Tutorial.ino
@@ -21,7 +21,7 @@ void setup()
 
 void loop() 
 {
-  byte myValue = 0;
+  int myValue = 0;
   myValue = analogRead(A0);
   
   Serial.print("The value is: ");
@@ -29,4 +29,3 @@ void loop()
 
   delay(250);
 }
-
